Move 'ensure local file' to project storage
