+++ AppBundles/ExportBOMPlugin/ExportBOMAutomation.cs
@@ -28,18 +28,6 @@ using Newtonsoft.Json.Serialization;
 
 namespace ExportBOMPlugin
 {
-    public class Column
-    {
-        public string Label { get; set; }
-        public bool? Numeric { get; set; }
-    }
-
-    public class ExtractedBOM
-    {
-        public Column[] Columns { get; set; }
-        public object[][] Data { get; set; }
-    }
-
     // heavily based on https://github.com/akenson/da-update-bom
     [ComVisible(true)]
     public class ExportBOMAutomation
