+++ AppBundles/ExportBOMPlugin/ExportBOMPlugin.csproj
@@ -47,6 +47,7 @@
     <Compile Include="..\..\Shared\SharedAssemblyInfo.cs">
       <Link>Properties\SharedAssemblyInfo.cs</Link>
     </Compile>
+    <Compile Include="ExtractedBOM.cs" />
     <Compile Include="PluginServer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ExportBOMAutomation.cs" />
