+++ AppBundles/ExportBOMPlugin/ExtractedBOM.cs
@@ -1,0 +1,14 @@
+﻿namespace ExportBOMPlugin
+{
+    public class Column
+    {
+        public string Label { get; set; }
+        public bool? Numeric { get; set; }
+    }
+
+    public class ExtractedBOM
+    {
+        public Column[] Columns { get; set; }
+        public object[][] Data { get; set; }
+    }
+}
