+++ WebApplication/State/UserResolver.cs
@@ -140,5 +140,32 @@ namespace WebApplication.State
 
             return Path.Combine(_localCache.LocalRootName, userDir);
         }
+
+        /// <summary>
+        /// Ensure that the file exists in local cache for a project.
+        /// </summary>
+        /// <param name="projectName">The project name.</param>
+        /// <param name="fileName">Short file name.</param>
+        /// <param name="hash">Parameters hash (default is used if not provided).</param>
+        /// <returns>Full path to the existing local file.</returns>
+        public async Task<string> EnsureLocalFile(string projectName, string fileName, string hash = null)
+        {
+            var projectStorage = await GetProjectStorageAsync(projectName);
+
+            var localNames = projectStorage.GetLocalNames(hash);
+            var fullLocalName = localNames.ToFullName(fileName);
+
+            if (! File.Exists(fullLocalName))
+            {
+                _logger.LogInformation($"Restoring missing '{fullLocalName}' for '{projectName}'");
+
+                Directory.CreateDirectory(localNames.BaseDir);
+
+                var bucket = await GetBucketAsync(tryToCreate: false);
+                await bucket.DownloadFileAsync(projectStorage.GetOssNames(hash).ToFullName(fileName), fullLocalName);
+            }
+
+            return fullLocalName;
+        }
     }
 }
